+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Fix a bug where edition screen was considered seen based on metadata props - maxim
 * Fix for selecting a painting in consignments - orta
+* Add auto focus to the work details / metadata screen - maxim
 
 ### 1.4.0-rc.0